Refactor ECS: Introduce ChangeFilter and Chunk Management
- Added ChangeFilter class to manage component change tracking.
- Implemented Chunk class to handle entity and component storage with versioning.
- Updated Archetype to support versioning and chunk management.
- Removed ChunkView class as its functionality is integrated into Chunk.
- Enhanced EntityMngr to utilize new chunk management and versioning features.
- Updated system and job management to respect component changes and versions.
- Refactored component access methods to include read and write access modes.
- Modified tests to align with new component access patterns and chunk management.

+++ include/MyECS/Entity.hpp
@@ -8,15 +8,15 @@
 namespace My::MyECS {
 // index + version
 struct Entity {
-  std::size_t index;
-  std::size_t version;
+  std::uint64_t index;
+  std::uint64_t version;
 
   static constexpr Entity Invalid() noexcept {
-    return {.index = static_cast<std::size_t>(-1),
-            .version = static_cast<std::size_t>(-1)};
+    return {.index = static_cast<std::uint64_t>(-1),
+            .version = static_cast<std::uint64_t>(-1)};
   }
   constexpr bool Valid() const noexcept {
-    return index != static_cast<std::size_t>(-1);
+    return index != static_cast<std::uint64_t>(-1);
   }
   constexpr auto operator<=>(const Entity&) const = default;
 };
